fix test warning
