+++ src/__tests__/useForm/formState.test.tsx
@@ -226,38 +226,6 @@ describe('formState', () => {
     });
   });
 
-  it('should bubble the error up when an error occurs in the provided handleSubmit function', async () => {
-    const errorMsg = 'this is an error';
-    const App = () => {
-      const [error, setError] = React.useState('');
-      const { register, handleSubmit } = useForm();
-
-      const rejectPromiseFn = jest.fn().mockRejectedValue(new Error(errorMsg));
-
-      return (
-        <form
-          onSubmit={() =>
-            handleSubmit(rejectPromiseFn)().catch((err) =>
-              setError(err.message),
-            )
-          }
-        >
-          <input {...register('test')} />
-          <p>{error}</p>
-          <button>Submit</button>
-        </form>
-      );
-    };
-
-    render(<App />);
-
-    await actComponent(async () => {
-      fireEvent.click(screen.getByRole('button'));
-    });
-
-    screen.getByText(errorMsg);
-  });
-
   it('should set isSubmitSuccessful to false when there is a promise reject', async () => {
     const App = () => {
       const {
